+++ tests/framework/db/QueryBuilderTest.php
@@ -1026,8 +1026,7 @@ abstract class QueryBuilderTest extends DatabaseTestCase
 
             [ ['in', 'id', [1]], '[[id]]=:qp0', [':qp0' => 1] ],
             [ ['in', 'id', new TraversableObject([1])], '[[id]]=:qp0', [':qp0' => 1] ],
-            // composite in
-            [
+            'composite in' => [
                 ['in', ['id', 'name'], [['id' =>1, 'name' => 'oy']]],
                 '([[id]], [[name]]) IN ((:qp0, :qp1))',
                 [':qp0' => 1, ':qp1' => 'oy']
@@ -1038,13 +1037,12 @@ abstract class QueryBuilderTest extends DatabaseTestCase
 
             [ ['in', 'id', new TraversableObject([1, 2, 3])], '[[id]] IN (:qp0, :qp1, :qp2)', [':qp0' => 1, ':qp1' => 2, ':qp2' => 3] ],
 
-            // composite in using array objects.
-            [
+            'composite in using array objects' => [
                 ['in', new TraversableObject(['id', 'name']), new TraversableObject([
                     ['id' => 1, 'name' => 'oy'],
                     ['id' => 2, 'name' => 'yo'],
                 ])],
-                '([[id]], [[name]]) IN ((:qp0, :qp1))',
+                '([[id]], [[name]]) IN ((:qp0, :qp1), (:qp2, :qp3))',
                 [':qp0' => 1, ':qp1' => 'oy', ':qp2' => 2, ':qp3' => 'yo']
             ],
             // exists