+++ framework/CHANGELOG.md
@@ -82,7 +82,7 @@ Yii Framework 2 Change Log
 - Eng #10976: `Inflector::transliterate()` now uses `strtr` instead of `str_replace` (DrDeath72)
 - Enh #11110: Added migrations for DB session (mdmunir)
 - Enh #10941: Added `yii\helpers\ArrayHelper::isTraversable`, added support for traversable selections for dropdownList, radioList and checkboxList in `yii\helpers\Html`.
-- Enh #10954: Query builder now supports `Traversable` objects for use in `in` conditions. (sammousa)
+- Enh #10954: `yii\db\QueryBuilder` now accepts `\Traversable` objects for `in` condition (SamMousa, silverfire)
 - Chg: HTMLPurifier dependency updated to `~4.6` (samdark)
 - Chg #9854: Added `ActiveRecordInterface::populateRelation()` to respect the methods called by the implementation (SamMousa)
 - Chg #10726: Added `yii\rbac\ManagerInterface::canAddChild()` (dkhlystov, samdark)